+++ mock_clusterClient_test.go
@@ -3,7 +3,7 @@
 package main
 
 import (
-	apiv1 "github.com/cloudogu/k8s-blueprint-lib/api/v1"
+	apiv2 "github.com/cloudogu/k8s-blueprint-lib/v2/api/v2"
 	apiserverinternalv1alpha1 "k8s.io/client-go/kubernetes/typed/apiserverinternal/v1alpha1"
 
 	appsv1 "k8s.io/client-go/kubernetes/typed/apps/v1"
@@ -2080,23 +2080,23 @@ func (_c *mockClusterClient_InternalV1alpha1_Call) RunAndReturn(run func() apise
 }
 
 // List provides a mock function with given fields: ctx, opts
-func (_m *mockClusterClient) List(ctx context.Context, opts metav1.ListOptions) (*apiv1.BlueprintList, error) {
+func (_m *mockClusterClient) List(ctx context.Context, opts metav1.ListOptions) (*apiv2.BlueprintList, error) {
 	ret := _m.Called(ctx, opts)
 
 	if len(ret) == 0 {
 		panic("no return value specified for List")
 	}
 
-	var r0 *apiv1.BlueprintList
+	var r0 *apiv2.BlueprintList
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, metav1.ListOptions) (*apiv1.BlueprintList, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, metav1.ListOptions) (*apiv2.BlueprintList, error)); ok {
 		return rf(ctx, opts)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, metav1.ListOptions) *apiv1.BlueprintList); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, metav1.ListOptions) *apiv2.BlueprintList); ok {
 		r0 = rf(ctx, opts)
 	} else {
 		if ret.Get(0) != nil {
-			r0 = ret.Get(0).(*apiv1.BlueprintList)
+			r0 = ret.Get(0).(*apiv2.BlueprintList)
 		}
 	}
 
@@ -2128,12 +2128,12 @@ func (_c *mockClusterClient_List_Call) Run(run func(ctx context.Context, opts me
 	return _c
 }
 
-func (_c *mockClusterClient_List_Call) Return(_a0 *apiv1.BlueprintList, _a1 error) *mockClusterClient_List_Call {
+func (_c *mockClusterClient_List_Call) Return(_a0 *apiv2.BlueprintList, _a1 error) *mockClusterClient_List_Call {
 	_c.Call.Return(_a0, _a1)
 	return _c
 }
 
-func (_c *mockClusterClient_List_Call) RunAndReturn(run func(context.Context, metav1.ListOptions) (*apiv1.BlueprintList, error)) *mockClusterClient_List_Call {
+func (_c *mockClusterClient_List_Call) RunAndReturn(run func(context.Context, metav1.ListOptions) (*apiv2.BlueprintList, error)) *mockClusterClient_List_Call {
 	_c.Call.Return(run)
 	return _c
 }