+++ packages/backup/service.go
@@ -2,11 +2,14 @@ package backup
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
+	"log/slog"
 	"time"
 
 	pbBackup "github.com/cloudogu/ces-control-api/generated/backup"
 	v1 "github.com/cloudogu/k8s-backup-lib/api/v1"
+	v2 "github.com/cloudogu/k8s-blueprint-lib/v2/api/v2"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
@@ -17,21 +20,34 @@ const (
 	backupStatusDeleting   = "deleting"
 )
 
+const (
+	blueprintIdAnnotation = "backup.cloudogu.com/blueprintId"
+	dogusAnnotation       = "backup.cloudogu.com/dogus"
+)
+
+// annotationDogus are found in the annotation "backup.cloudogu.com/dogus" of a backup
+type annotationDogus struct {
+	Name    string `json:"name"`
+	Version string `json:"version"`
+}
+
 type DefaultBackupService struct {
 	pbBackup.UnimplementedBackupManagementServer
 	backupClient         backupInterface
 	restoreClient        restoreInterface
 	backupScheduleClient backupScheduleClient
 	componentClient      componentClient
+	blueprintLister      blueprintLister
 }
 
 // NewBackupService returns an instance of defaultBackupService.
-func NewBackupService(backupClient backupInterface, restoreClient restoreInterface, backupScheduleClient backupScheduleClient, componentClient componentClient) *DefaultBackupService {
+func NewBackupService(backupClient backupInterface, restoreClient restoreInterface, backupScheduleClient backupScheduleClient, componentClient componentClient, blueprintLister blueprintLister) *DefaultBackupService {
 	return &DefaultBackupService{
 		backupClient:         backupClient,
 		restoreClient:        restoreClient,
 		backupScheduleClient: backupScheduleClient,
 		componentClient:      componentClient,
+		blueprintLister:      blueprintLister,
 	}
 }
 
@@ -68,7 +84,61 @@ func (s *DefaultBackupService) AllBackups(ctx context.Context, _ *pbBackup.GetAl
 		return nil, fmt.Errorf("failed to list backups: %w", err)
 	}
 
-	return &pbBackup.GetAllBackupsResponse{Backups: s.mapBackups(list)}, nil
+	blueprintList, err := s.blueprintLister.List(ctx, metav1.ListOptions{})
+	if err != nil {
+		return nil, fmt.Errorf("failed to get blueprint: %w", err)
+	}
+	if len(blueprintList.Items) == 0 {
+		return nil, fmt.Errorf("failed to get blueprint: no blueprints available")
+	}
+
+	backups := s.mapBackups(list, &blueprintList.Items[0])
+	return &pbBackup.GetAllBackupsResponse{Backups: backups}, nil
+}
+
+// CreateRestore creates a restore for the given backup.
+// The restore is only created if the backup is restorable.
+func (s *DefaultBackupService) CreateRestore(ctx context.Context, request *pbBackup.CreateRestoreRequest) (*pbBackup.CreateRestoreResponse, error) {
+	backup, err := s.backupClient.Get(ctx, request.BackupId, metav1.GetOptions{})
+	if err != nil {
+		return nil, fmt.Errorf("failed to get backup: %w", err)
+	}
+
+	list, err := s.blueprintLister.List(ctx, metav1.ListOptions{})
+	if err != nil {
+		return nil, fmt.Errorf("failed to get blueprint: %w", err)
+	}
+	if len(list.Items) == 0 {
+		return nil, fmt.Errorf("failed to get blueprint: blueprint not found")
+	}
+
+	// only create restore if the backup is restorable
+	restorable, err := s.isBackupRestorable(backup, &list.Items[0])
+	if err != nil {
+		return nil, fmt.Errorf("failed to check if backup is restorable: %w", err)
+	}
+	if restorable {
+		timestamp := time.Now().Format("20060102-1504")
+		restoreName := fmt.Sprintf("restore-%s", timestamp)
+		restore := &v1.Restore{
+			TypeMeta: metav1.TypeMeta{},
+			ObjectMeta: metav1.ObjectMeta{
+				Name: restoreName,
+			},
+			Spec: v1.RestoreSpec{
+				BackupName: request.BackupId,
+			},
+			Status: v1.RestoreStatus{},
+		}
+		_, err := s.restoreClient.Create(ctx, restore, metav1.CreateOptions{})
+		if err != nil {
+			return nil, fmt.Errorf("failed to create restore: %w", err)
+		}
+	} else {
+		return nil, fmt.Errorf("backup is not restorable")
+	}
+
+	return &pbBackup.CreateRestoreResponse{}, nil
 }
 
 func (s *DefaultBackupService) AllRestores(ctx context.Context, _ *pbBackup.GetAllRestoresRequest) (*pbBackup.GetAllRestoresResponse, error) {
@@ -85,12 +155,18 @@ func (s *DefaultBackupService) AllRestores(ctx context.Context, _ *pbBackup.GetA
 	return &pbBackup.GetAllRestoresResponse{Restores: restores}, nil
 }
 
-func (s *DefaultBackupService) mapBackups(backupList *v1.BackupList) []*pbBackup.BackupResponse {
+func (s *DefaultBackupService) mapBackups(backupList *v1.BackupList, blueprint *v2.Blueprint) []*pbBackup.BackupResponse {
 	backupResponseList := make([]*pbBackup.BackupResponse, 0, 5)
 	for _, backup := range backupList.Items {
 		// skip backups in deleting state
 		if backup.Status.Status == backupStatusDeleting {
 			continue
+		}
+		restorable, err := s.isBackupRestorable(&backup, blueprint)
+		if err != nil {
+			// There might be backups that do not have the annotations. In this case we just log the error and continue.
+			slog.Error(fmt.Sprintf("failed to check if backup is restorable: %v", err))
+			restorable = false
 		}
 		backupResponse := pbBackup.BackupResponse{
 			Id:             backup.Name,
@@ -98,6 +174,8 @@ func (s *DefaultBackupService) mapBackups(backupList *v1.BackupList) []*pbBackup
 			EndTime:        backup.Status.CompletionTimestamp.String(),
 			Status:         backupStatus(&backup),
 			CurrentVersion: true,
+			Restorable:     restorable && backup.Status.Status == backupStatusCompleted,
+			BlueprintId:    backup.GetAnnotations()[blueprintIdAnnotation],
 		}
 		backupResponseList = append(backupResponseList, &backupResponse)
 	}
@@ -118,7 +196,7 @@ func (s *DefaultBackupService) mapRestores(ctx context.Context, restoreList *v1.
 			StartTime:   backup.Status.StartTimestamp.String(),
 			EndTime:     backup.Status.CompletionTimestamp.String(),
 			Success:     restore.Status.Status == "completed",
-			BlueprintId: "Unknown", // FIXME: fix once backup POC is completed and BlueprintId has been added in some shape or form.
+			BlueprintId: backup.GetAnnotations()[blueprintIdAnnotation],
 		}
 		restoreResponseList = append(restoreResponseList, &restoreResponse)
 	}
@@ -166,6 +244,47 @@ func (s *DefaultBackupService) GetRetentionPolicy(ctx context.Context, _ *pbBack
 	}
 
 	return &pbBackup.GetRetentionPolicyResponse{Policy: retentionPolicy}, nil
+}
+
+// a backup is restorable if it is from the same blueprint and the dogus are matching
+func (s *DefaultBackupService) isBackupRestorable(backup *v1.Backup, blueprint *v2.Blueprint) (bool, error) {
+	ans := backup.GetAnnotations()
+
+	// get all dogus from backup annotations
+	backupDogus := make([]annotationDogus, 0, 5)
+	err := json.Unmarshal([]byte(ans[dogusAnnotation]), &backupDogus)
+	if err != nil {
+		return false, fmt.Errorf("failed to unmarshal dogus from backup: %w", err)
+	}
+
+	blueprintNameIsMatching := ans != nil && ans[blueprintIdAnnotation] == blueprint.Spec.DisplayName
+
+	if blueprintNameIsMatching && s.isDoguListMatching(backupDogus, blueprint.Spec.Blueprint.Dogus) {
+		return true, nil
+	}
+	return false, nil
+}
+
+// isDoguListMatching checks if the given list of backup dogus is matching the given list of dogus from the blueprint.
+func (s *DefaultBackupService) isDoguListMatching(backupDogus []annotationDogus, blueprintDogus []v2.Dogu) bool {
+	if len(backupDogus) != len(blueprintDogus) {
+		return false
+	}
+
+	backupDoguMap := make(map[string]annotationDogus)
+	for _, v := range backupDogus {
+		backupDoguMap[v.Name] = v
+	}
+	for _, blueprintDogu := range blueprintDogus {
+		backupDogu, ok := backupDoguMap[blueprintDogu.Name]
+		if !ok {
+			return false
+		}
+		if backupDogu.Version != *blueprintDogu.Version {
+			return false
+		}
+	}
+	return true
 }
 
 func backupStatus(backup *v1.Backup) string {